+++ src/compl/mod.rs
@@ -5,5 +5,5 @@
 //!
 //! [clap's docs.rs website]: https://docs.rs/clap/2.33.3/clap/enum.Shell.html
 
-pub mod compl_arg;
-pub mod compl_handler;
+pub mod compl_args;
+pub mod compl_handlers;