+++ src/config/deserialized_account_config.rs
@@ -1,6 +1,8 @@
 use serde::Deserialize;
 use std::{collections::HashMap, path::PathBuf};
 
+use crate::config::Format;
+
 pub trait ToDeserializedBaseAccountConfig {
     fn to_base(&self) -> DeserializedBaseAccountConfig;
 }
@@ -9,18 +11,22 @@ pub trait ToDeserializedBaseAccountConfig {
 #[derive(Debug, Clone, Deserialize)]
 #[serde(untagged)]
 pub enum DeserializedAccountConfig {
+    #[cfg(feature = "imap-backend")]
     Imap(DeserializedImapAccountConfig),
+    #[cfg(feature = "maildir-backend")]
     Maildir(DeserializedMaildirAccountConfig),
-    #[cfg(feature = "notmuch")]
+    #[cfg(feature = "notmuch-backend")]
     Notmuch(DeserializedNotmuchAccountConfig),
 }
 
 impl ToDeserializedBaseAccountConfig for DeserializedAccountConfig {
     fn to_base(&self) -> DeserializedBaseAccountConfig {
         match self {
+            #[cfg(feature = "imap-backend")]
             Self::Imap(config) => config.to_base(),
+            #[cfg(feature = "maildir-backend")]
             Self::Maildir(config) => config.to_base(),
-            #[cfg(feature = "notmuch")]
+            #[cfg(feature = "notmuch-backend")]
             Self::Notmuch(config) => config.to_base(),
         }
     }
@@ -47,6 +53,9 @@ macro_rules! make_account_config {
             pub notify_query: Option<String>,
             /// Overrides the watch commands for this account.
             pub watch_cmds: Option<Vec<String>>,
+	    /// Represents the text/plain format as defined in the
+	    /// [RFC2646](https://www.ietf.org/rfc/rfc2646.txt)
+            pub format: Option<Format>,
 
             /// Makes this account the default one.
             pub default: Option<bool>,
@@ -89,6 +98,7 @@ macro_rules! make_account_config {
             	    notify_cmd: self.notify_cmd.clone(),
             	    notify_query: self.notify_query.clone(),
             	    watch_cmds: self.watch_cmds.clone(),
+            	    format: self.format.clone(),
 
             	    default: self.default.clone(),
             	    email: self.email.clone(),
@@ -112,6 +122,7 @@ macro_rules! make_account_config {
 
 make_account_config!(DeserializedBaseAccountConfig,);
 
+#[cfg(feature = "imap-backend")]
 make_account_config!(
     DeserializedImapAccountConfig,
     imap_host: String,
@@ -122,9 +133,10 @@ make_account_config!(
     imap_passwd_cmd: String
 );
 
+#[cfg(feature = "maildir-backend")]
 make_account_config!(DeserializedMaildirAccountConfig, maildir_dir: String);
 
-#[cfg(feature = "notmuch")]
+#[cfg(feature = "notmuch-backend")]
 make_account_config!(
     DeserializedNotmuchAccountConfig,
     notmuch_database_dir: String