+++ src/output/print_table.rs
@@ -2,6 +2,8 @@ use anyhow::Result;
 use std::io;
 use termcolor::{self, StandardStream};
 
+use crate::config::Format;
+
 pub trait WriteColor: io::Write + termcolor::WriteColor {}
 
 impl WriteColor for StandardStream {}
@@ -10,6 +12,7 @@ pub trait PrintTable {
     fn print_table(&self, writter: &mut dyn WriteColor, opts: PrintTableOpts) -> Result<()>;
 }
 
-pub struct PrintTableOpts {
+pub struct PrintTableOpts<'a> {
+    pub format: &'a Format,
     pub max_width: Option<usize>,
 }